feat: implement backend GraphQL schema and resolver updates - Task 1.0 Complete - Backend GraphQL Schema and Resolver Updates with previewScan query, ScanPreview object type, authentication for createNewScan, comprehensive unit tests, and frontend GraphQL type fixes

+++ frontend/src/graphql/mutations.ts
@@ -19,7 +19,7 @@ mutation CreateNewScan($data: ScanInput!) {
   }`
 
 export const DELETE_SCAN = gql`
-mutation DeleteScan($deleteScanId: Float!) {
+mutation DeleteScan($deleteScanId: Int!) {
     deleteScan(id: $deleteScanId)
   }`
