+++ frontend/src/graphql/queries.ts
@@ -20,7 +20,7 @@ query GetAllScans {
 }`
 
 export const GET_SCAN_BY_ID = gql`
-query Query($getScanByIdId: Float!) {
+query Query($getScanByIdId: Int!) {
   getScanById(id: $getScanByIdId) {
     id
     url
